backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10
Flask-Cors==5.0.0
Werkzeug==3.1.3
python-dotenv==1.0.1
PyJWT==2.9.0